+++ input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gluestack-ui/input
 
+## 0.1.6
+
+### Patch Changes
+
+- feat: migration of hook to react native aria
+- Updated dependencies
+  - @gluestack-ui/form-control@0.1.7
+
 ## 0.1.5
 
 ### Patch Changes